RHTAPSRE-458: Add admission webhook
The admission webhook will validate the creation and update of Repository
objects. It will enforce that the url of the repository matches one
of the prefixes provided in a config file.

Using this admission webhook, an administrator can control which
repositories can be used with pipelines-as-code.

The core logic and tests are included under the `api` package.

This project was scaffolded using kubebuilder.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>